# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Dollar CARE'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  /api/v1/doctor/changePassword:
    post:
      summary: 'Change Password'
      operationId: changePassword
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n\"status\": true,\n\"statusCode\": 200,\n\"message\": \"Password changed successfully\",\n\"data\": {\n    \"specializations\": [\n        {\n            \"id\": 2,\n            \"name\": \"Dentist\",\n            \"slug\": \"dentist\",\n            \"image\": \"specialization/1dyavyARmT4SiXK9izxsvF1vjW0PYXI28mB94TuB.png\",\n            \"description\": \"Visit your doctor for joint pain, sprains, arthritis, and other bone pains.\",\n            \"status\": 1,\n            \"created_at\": \"2023-06-06T12:40:19.000000Z\",\n            \"updated_at\": \"2023-06-14T08:51:04.000000Z\"\n        },\n    ],\n    \"user\": {\n            \"id\": 4,\n            \"name\": \"James Bond\",\n            \"email\": \"james@yopmail.com\",\n            \"phone\": \"08596769586\",\n            \"gender\": \"Male\",\n            \"age\": \"2001-01-30\",\n            \"license_number\": \"UPS74856963\",\n            \"profile_picture\": \"doctor/5HiPk9oN9cQCNNKdzmBgvNLHSL8u7bbHncdrPE91.png\",\n            \"fcm_token\": \"eyo_Z9a_EQkcwF1y_TjA-U:APA91bFvEyoQw3XaMNheNcHlQMQISc-sB2b3RWySS515RaMauTX1RCKrinBcf5-4FhZeLtoB-BYV4K0khjO8pD1ZTi257s3A6_sH36SZGN4tBdS1AlUo3L2sBaXKpZoI0gwq-MWETDfN\",\n            \"status\": 1,\n            \"specializations\": [\n                {\n                    \"id\": 4,\n                    \"name\": \"ENT Specialist\",\n                    \"pivot\": {\n                        \"doctor_id\": 4,\n                        \"specialization_id\": 4\n                    }\n                }\n            ]\n    }\n}\n}"
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'The old_password and password must be different.'
                properties:
                  error:
                    type: string
                    example: 'The old_password and password must be different.'
      tags:
        - 'Doctor Profile'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                  description: 'The old_password of the Doctor.'
                  example: odio
                new_password:
                  type: string
                  description: 'The password of the Doctor.'
                  example: dolorem
                confirm_password:
                  type: string
                  description: 'The password_confirmation of the Doctor.'
                  example: harum
              required:
                - current_password
                - new_password
                - confirm_password
...
  /api/v1/patient/changePassword:
    post:
      summary: 'Change Password'
      operationId: changePassword
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  statusCode: 200
                  message: 'Password changed successfully'
                properties:
                  status:
                    type: boolean
                    example: true
                  statusCode:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: 'Password changed successfully'
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'The old_password and password must be different.'
                properties:
                  error:
                    type: string
                    example: 'The old_password and password must be different.'
      tags:
        - 'Patient Profile'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                  description: 'The old_password of the Patient.'
                  example: dolorum
                new_password:
                  type: string
                  description: 'The password of the Patient.'
                  example: fugit
                confirm_password:
                  type: string
                  description: 'The password_confirmation of the Patient.'
                  example: temporibus
              required:
                - current_password
                - new_password
                - confirm_password
...